+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.19.0] - 2023-08-31
+
+### Added
+- Added official support for the latest LTS Editor 2022.3 while maintaining support for 2021.3.
+
+### Changed
+- Default `PkceConfiguration` TokenUrl, RefreshTokenUrl and LogoutUrl no longer use cloud service endpoints.
+- [Breaking] WebGL platform use secure document.cookie instead of HTTP headers to reach cloud service endpoints.
+
 ## [0.18.1] - 2023-08-17
 
 ### Fixed